Switch to new bullet system
Rename bIsSpacePressed to bReleased
Implement first bullet wave for demon

+++ FlockingHell/FlockingHell/Source/Public/CirclePattern.h
@@ -15,7 +15,7 @@ struct CirclePattern : BulletPatternGenerator
 	void Update() override;
 	void Draw() override;
 
-	void Delay(unsigned short Seconds) override;
+	void Delay(float Seconds) override;
 
 	void AddDebugSwitchPatternCode() override;
 	void DrawDebugInfo() override;
