+++ FlockingHell/FlockingHell/Source/Public/SpreadPattern.h
@@ -15,7 +15,7 @@ struct SpreadPattern : BulletPatternGenerator
 	void Update() override;
 	void Draw() override;
 
-	void Delay(unsigned short Seconds) override;
+	void Delay(float Seconds) override;
 
 	void AddDebugSwitchPatternCode() override;
 	void DrawDebugInfo() override;
